price type 변경

+++ src/main/java/com/lokoko/domain/product/application/crawler/NewProductCrawler.java
@@ -10,6 +10,8 @@ import com.lokoko.domain.product.domain.entity.enums.Tag;
 import com.lokoko.domain.product.domain.repository.ProductRepository;
 import com.lokoko.global.utils.ProductCrawlerConstants;
 import com.lokoko.global.utils.ProductCrawlerUtil;
+
+import java.math.BigDecimal;
 import java.time.Duration;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -124,7 +126,7 @@ public class NewProductCrawler {
             }
 
             Product p = Product.builder()
-                    .normalPrice(price)
+                    .normalPrice(BigDecimal.valueOf(price))
                     .productName(detail)
                     .shippingInfo(ship != null ? ship : "배송 정보 없음")
                     .tag(tag)
